Refactor code indexer
